Sony T1: Support for copying covers similar to the 505 driver.
Performs the copy during sync_booklists due to having access to the book's id/lpath at that point, and being able to detect new books. When upload_cover is normally called, this information is usually not accessible, and the book isn't actually in the database. It's less fragile this way.

Fixes an issue with setting floating point values in 'added_time' column.
Also removes Book from books.py (not needed at this point)